+++ man/examples/scr_LSR.R
@@ -1,23 +1,43 @@
 # Example on loading
 # long-short ratio
-# for the last 5 days
+# for the last days
 # on the 15 minute candle
 # wrapped in try to avoid
 # failure on Github
 BTC_LSR <- try(
-  getLSRatio(
+  cryptoQuotes::getLSRatio(
     ticker = 'BTCUSDT',
     interval = '15m',
-    from = Sys.Date() - 5,
+    from = Sys.Date() - 1,
     to   = Sys.Date()
   )
 )
 
-# head the data;
+
+BTCUSDT <- try(
+  cryptoQuotes::getQuote(
+    ticker = 'BTCUSDT',
+    futures = TRUE,
+    interval = '15m',
+    from = Sys.Date() - 1,
+    to   = Sys.Date()
+  )
+)
+
+
+# plot and head the data
 if (!inherits(x = BTC_LSR, what = 'try-error')) {
 
   head(
     BTC_LSR
+  )
+
+  cryptoQuotes::chart(
+    chart = cryptoQuotes::kline(
+      BTCUSDT
+    ) %>% cryptoQuotes::addLSRatio(
+      LSR = BTC_LSR
+    )
   )
 
 }